REFACTOR: Using generic function to check if material is associated with skill
