+++ components/Tags/SkillTag.tsx
@@ -1,9 +1,6 @@
 "use client";
 
-import hasAssociatedSkills from "@/actions/skills/hasAssociatedSkills";
-import isSkillAssociatedWithBlogs from "@/actions/skills/isSkillAssociatedWithBlogs";
-import isSkillAssociatedWithCertificate from "@/actions/skills/isSkillAssociatedWithCertificate";
-import isSkillAssociatedWithProject from "@/actions/skills/isSkillAssociatedWithProject";
+import isSkillAssociatedWithMaterial from "@/actions/material/isSkillAssociatedWithMaterial";
 import {
   Tooltip,
   TooltipContent,
@@ -13,6 +10,9 @@ import blogs from "@/database/blogs";
 import certificates from "@/database/certificates";
 import allProjects from "@/database/projects";
 import allSkills from "@/database/skills/skills";
+import BlogInterface from "@/interfaces/material/BlogInterface";
+import MaterialInterface from "@/interfaces/material/MaterialInterface";
+import ProjectInterface from "@/interfaces/material/ProjectInterface";
 import SkillInterface from "@/interfaces/skills/SkillInterface";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
@@ -27,15 +27,12 @@ interface TagProps {
 const SkillTag: React.FC<TagProps> = ({ skill, hide }) => {
   const currentPath = usePathname();
 
-  const skills = allSkills;
-  const allBlogs = blogs;
-  const projects = allProjects;
+  const skills: SkillInterface[] = allSkills;
+  const allBlogs: BlogInterface[] = blogs;
+  const projects: ProjectInterface[] = allProjects;
+  const allMaterial: MaterialInterface[] = [...projects, ...certificates, ...allBlogs];
 
-  const hasProjects = isSkillAssociatedWithProject(skill, projects);
-  const hasCertificates = isSkillAssociatedWithCertificate(skill, certificates);
-  const hasBlogs = isSkillAssociatedWithBlogs(skill, allBlogs);
-  const hasSkill = hasAssociatedSkills(skill, skills);
-  const hasMaterial = hasProjects || hasCertificates || hasBlogs;
+  const hasMaterial = isSkillAssociatedWithMaterial(skill, allMaterial);
 
   if (hide && !hasMaterial) {
     return <></>;
